IMPALA-9690: Set minimum CPU requirement to AVX2 for x86_64
AVX2 has been available on x86_64 processors for several
years. To simplify the code going forward, this
changes Impala to require AVX2 for x86_64. If AVX2
is not present, Impala will refuse to start up.

Prior to this change, the LLVM IR code for x86_64 was
built with two versions: one that requires SSE4 and
another that does not. This simplifies the code to
have only one version that is always built with AVX2
support.

This does not have any impact on ARM support.

Testing:
 - Core job passes
 - Release exhaustive job passes
 - Performance run on TPC-H scale 42 shows minor improvement.
+----------+-----------------------+---------+------------+------------+----------------+
| Workload | File Format           | Avg (s) | Delta(Avg) | GeoMean(s) | Delta(GeoMean) |
+----------+-----------------------+---------+------------+------------+----------------+
| TPCH(42) | parquet / none / none | 8.38    | -1.02%     | 5.28       | -1.01%         |
+----------+-----------------------+---------+------------+------------+----------------+

Change-Id: I6de2e897d8ff24dea3c7c1c4ab9d61734839835d
Reviewed-on: http://gerrit.cloudera.org:8080/17406
Reviewed-by: Jim Apple <jbapple@apache.org>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>