IMPALA-7556: Decouple BufferManagement from the ScanRange and IoMgr
Currently BufferManagement is tightly coupled with ScanRange.
Every ScanRange maintains list of unused buffers and ready buffers.
Unused buffers are buffers used to read scanned data and ready
buffers are buffers with the data already read. For managing these
buffers, ScanRange defines various functions like AddUnusedBuffer,
GetUsedBuffer, EnqueueReadyBuffer and functions to allocate and
cleanup buffers. This patch has created ScanBufferManager which
would be responsible for the managing these buffers for ScanRange.
ScanBufferManager's logic is still coupled with the ScanRange,
but refactorig it into a seperate class is a good first step.

Testing:
 1. Ran these existing tests: EE, BackEnd, JDBC and Cluster test.
 2. Ran the above tests with TSAN build using exhaustive strategy.

Change-Id: Ibd74691b50b46114f95a8641034c05d07ddeec97
Reviewed-on: http://gerrit.cloudera.org:8080/17413
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>