IMPALA-10694: Improve the error handling of setup-ranger
We found that setup-ranger will continue the execution if an error
occurs when i) wget was executed to initialize the environment
variables of GROUP_ID_OWNER and GROUP_ID_NON_OWNER, and ii) curl was
executed to upload a revised Ranger policy even though the -e option was
set when create-load-data.sh was executed. This patch improves the error
handling by making setup-ranger exit as soon as an error occurs so that
no test would be run at all in case there is an error.

To exit if an error occurs during wget, we separate the assignment and
the export of the environment variables since the export of an
environment variable will run and succeed even though there is an error
before the export if the assignment and the export are combined. That
is, combining them hides the error.

On the other hand, to exit if an error occurs during curl, we add an
additional -f option so that an error will no longer be silently
ignored.

Testing:
 - Verified that setup-ranger could be successfully executed after this
   patch.
 - Verified that setup-ranger would exit if a URL in setup-ranger is not
   correctly set up or if the 'id' field in policy_4_revised.json does
   not match the URL of the policy to be updated.

Change-Id: I45605d1a7441b734cf80249626638cde3adce28b
Reviewed-on: http://gerrit.cloudera.org:8080/17386
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>