IMPALA-7844: Default to not allow ordinals in HAVING clause
Base on the discussion of the previous patch, we decide to not allow
ordinals in the HAVING clause since 4.0. It's a non-standard feature
that unintentionally supported by Impala 3.x and earlier versions.

This patch disables it by default, and add a feature flag to turn it on
for users that do depend on it.

Tests:
 - Modify existing FE tests to test on the flag.
 - Add custom cluster test to verify the flag works.

Change-Id: I0a57b8b65b046fae483e485e8391f8222fa586a5
Reviewed-on: http://gerrit.cloudera.org:8080/17415
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>