Bump up GBN to 13088653
This change bumps up the CDP_BUILD_NUMBER to 13088653
which includes additional metastore thrift API changes
needed for Catalogd's metastore endpoint to be functional.

Additionally, this adds a exclusion in the ranger-audit-plugin
dependency because it leaks hive-exec jar in the classpath which
causes thrift errors during compilation. The compilation error occur
because hive-exec leaks 0.9.3 thrift while impala now uses 0.11.0
version of thrift.

Testing:
1. Verified that code builds locally.
2. Run core tests.

Change-Id: Ie3ad6311f0ef1a06391b70d58cdb4e51e08de5ea
Reviewed-on: http://gerrit.cloudera.org:8080/17424
Reviewed-by: Fang-Yu Rao <fangyu.rao@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>