IMPALA-10702: Add warning logs for slow or large catalogd response
It'd be helpful to log the slow or large responses of catalogd in
debugging scalability issues. This patch adds these warning logs in
JniCatalog, where we serialize thrift responses. See some example
outputs in the jira description.

Responses that have size larger than 50MB or take more than 60s to
finish will be logged with the request. Add flags for these two
thredshold in case users found the warnings too verbose and want to
increase the thresholds.

Change-Id: Icffcfcaad2a718aebf79e2331efb05ca7a9a7671
Reviewed-on: http://gerrit.cloudera.org:8080/17427
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>