IMPALA-10197 (Part 2): Add KUDU_REPLICA_SELECTION query option
The previous patch added a new test case in PlannerTest for Kudu,
which check for specific number of hosts and instances for distributed
plan with query option KUDU_REPLICA_SELECTION set as LEADER_ONLY.
However, the leadership isn't deterministic since each Kudu partition
has its own Raft group and the leaders of these groups are not exactly
balanced across hosts in a cluster. There's no guarantee that we'll get
a certain number of leader hosts for a query which access multiple KUDU
partitions. This makes the unit-test flaky.

This patch removed the distributed plans to avoid nondeterminism.

Testing:
 - Reran the Planner test and verified the issue did not happen.

Change-Id: I3e23667c06c273a261e03de3d81fc7ee1f6b0682
Reviewed-on: http://gerrit.cloudera.org:8080/17502
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>