IMPALA-10187: Event processing fails on multiple events + DROP TABLE
Event processing can fail when it needs to process multiple events
belonging to the same table, and if one of the events is a DROP
TABLE.

E.g. imagine the following scenario:

Hive> alter table foo set location '/tmp';
Hive> drop table foo;

If the two DDLs are executed quickly there's good chance that Impala's
event processor will handle these in one event batch. In this case
during the processing of ALTER TABLE Impala will try to load table
'foo' which will fail (because Hive already deleted the table).

The event processor is not prepared for this exception and gets into
a broken state which means no further events getting processed.

This patch adds exception handling to
CatalogServiceCatalog.reloadTableIfExists() for the case when the
table cannot be loaded, in which case it returns false.

Testing:
 * added e2e test

Change-Id: I2bd72247293d830063f66eecf9e4ba69a439a605
Reviewed-on: http://gerrit.cloudera.org:8080/17542
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>