IMPALA-10711: Allow customizing startup's filesystem check
At startup, Impala runs JniFrontend::checkFileSystem() to verify
that it can list the root directory of the filesystem. Some
deployments would like to restrict Impala to a subdirectory
of the filesystem. In that circumstance, Impala may not have
access to the root of the filesystem and the startup check
will fail.

This adds the startup parameter: startup_filesystem_check_directories
to specify the directories to check in JniFrontend::checkFileSystem()
at startup. For deployments that are restricted to a subdirectory,
startup_filesystem_check_directories=/some/subdirectory would
check /some/subdirectory rather than /. This parameter allows for
a list of directories in case it is useful to verify multiple
locations (such as different s3 buckets). If set to the empty string,
no directories will be checked.

Testing:
 - Added test_startup_filesystem_checks.py custom cluster test
 - Hand tested locally

Change-Id: I68e6802424f16002c90a37b3ae2490721525e58d
Reviewed-on: http://gerrit.cloudera.org:8080/17570
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>