IMPALA-9355: TestExchangeMemUsage.test_exchange_mem_usage_scaling doesn't hit the memory limit
This patch restores the previous testing enviroment in which exchange
node 5 in the test query shown below reliablly reports memory
exceeding limit error.

  set mem_limit=171520k;
  set num_scanner_threads=1;
  select *
  from tpch_parquet.lineitem l1
    join tpch_parquet.lineitem l2 on l1.l_orderkey = l2.l_orderkey and
        l1.l_partkey = l2.l_partkey and l1.l_suppkey = l2.l_suppkey
        and l1.l_linenumber = l2.l_linenumber
  order by l1.l_orderkey desc, l1.l_partkey, l1.l_suppkey, l1.l_linenumber
  limit 5

In the patch, the memory limit for the entire query is lowered to 164Mb,
less than 171520k, the minimum for the query to be accepted after
the result spooling is enabled by default. This is achieved by
disabling result spooling by setting spool_query_results to false.

Testing:
  1). Ran exchange-mem-scaling.test

Change-Id: Ia4ad4508028645b67de419cfdfa2327d2847cfc4
Reviewed-on: http://gerrit.cloudera.org:8080/17586
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>